Moved hex_to_bytes fn to utils.
