+++ bittensor/core/subtensor.py
@@ -71,7 +71,13 @@ from bittensor.core.extrinsics.transfer import (
     transfer_extrinsic,
 )
 from bittensor.core.metagraph import Metagraph
-from bittensor.utils import networking, torch, ss58_to_vec_u8, u16_normalized_float
+from bittensor.utils import (
+    networking,
+    torch,
+    ss58_to_vec_u8,
+    u16_normalized_float,
+    hex_to_bytes,
+)
 from bittensor.utils.balance import Balance
 from bittensor.utils.btlogging import logging
 from bittensor.utils.registration import legacy_torch_api_compat
@@ -83,14 +89,6 @@ KEY_NONCE: dict[str, int] = {}
 class ParamWithTypes(TypedDict):
     name: str  # Name of the parameter.
     type: str  # ScaleType string of the parameter.
-
-
-def hex_to_bytes(hex_str: str) -> bytes:
-    try:
-        bytes_result = bytes.fromhex(hex_str[2:])
-    except ValueError:
-        bytes_result = bytes.fromhex(hex_str)
-    return bytes_result
 
 
 class Subtensor:
